+++ src/Controller/SearchController.php
@@ -4,21 +4,23 @@
 namespace App\Controller;
 
 use App\Form\SearchBarType;
-use App\Repository\SearchRepository;
+use App\Repository\PlaceRepository;
+use App\Repository\UserRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 
 
-
 class SearchController extends Controller
 {
     /**
-     * 
+     * @Route("/search", name="search")
      */
-    public function searchPaxAndPlace(Request $request, SearchRepository $searchRepo) {
+    public function searchPaxAndPlace(Request $request, UserRepository $userRepo, PlaceRepository $placeRepo) {
        
-       $form = $this->createForm(SearchBarType::class);
+       $form = $this->createForm(SearchBarType::class, [
+           'attr' => ['class' => 'h-25 d-inline-block mw-100']
+       ]);
        
        $form->handleRequest($request);
        
@@ -26,18 +28,21 @@ class SearchController extends Controller
        
         if ($form->isSubmitted() && $form->isValid()) {
            
-            $datas = $form->getData();
+            $data= $form->getData();
             
-            foreach ($datas as $data) {
-               $results = $searchRepo->findByPax($data); 
+            if($data['Recherche'] == 'username') {
+                $results = $userRepo->findByPax($data['search']);
+            } else {
+                $results = $placeRepo->findByPlace($data['search']);
             }
-            
-            // $this->json($results);
+             
+            return $this->render('search/searchResult.html.twig', [
+                'results' => $results
+            ]);
          }
     
          return $this->render('search/searchBar.html.twig', [
-             'form' => $form->createView(),
-             'results' => $results
+             'form' => $form->createView()
              ]);
          
     }
