+++ src/Repository/UserRepository.php
@@ -24,7 +24,19 @@ class UserRepository extends ServiceEntityRepository implements UserLoaderInterf
             ->getOneOrNullResult();
     }
     
-    
+    public function findByPax($username) 
+    {
+                $qb = $this->createQueryBuilder('u');
+                
+                $qb
+                    ->select('u.username, u.id, u.photoProfile, u.description')
+                    ->where('u.username LIKE :username')->setParameter(':username', '%'.$username.'%')
+                    ->orderBy('u.username', 'ASC');
+                
+        
+        $results = $qb->getQuery()->getResult();
+        return $results;
+    }
     
     /*
     public function findBySomething($value)
