+++ templates/base.html.twig
@@ -83,7 +83,7 @@
             
         </header>
         
-        <main class="container">
+        <main class="container text-center">
             
             {% include 'modal_base.html.twig' %}
