+++ templates/home.html.twig
@@ -4,7 +4,6 @@
 
 {% block body %}
 
-    {{render(controller(
-            'App\\Controller\\SearchController::searchPaxAndPlace', { 'request' : app.request } )) }} 
+     
 
 {% endblock %}
